Move container code to component file
