+++ client/src/components/Admin/Users/Filter.js
@@ -1,4 +1,6 @@
 import React, { PropTypes } from 'react';
+import { connect } from 'react-redux';
+import setUserFilter from '../../../actionCreators/setUserFilter';
 
 const UserFilter = ({ filter, onChange }) => (
   <div className="UserFilter">
@@ -17,4 +19,18 @@ UserFilter.propTypes = {
   onChange: PropTypes.func.isRequired,
 };
 
+const mapStateToProps = state => ({
+  filter: state.userFilter,
+});
+
+const mapDispatchToProps = dispatch => ({
+  onChange: (filter) => {
+    dispatch(setUserFilter(filter));
+  },
+});
+
 export default UserFilter;
+export const UserFilterContainer = connect(
+  mapStateToProps,
+  mapDispatchToProps,
+)(UserFilter);
