+++ client/src/components/Admin/Users/List.js
@@ -1,4 +1,7 @@
 import React, { PropTypes } from 'react';
+import { connect } from 'react-redux';
+import Fuse from 'fuse.js';
+import { toggleCanVote } from '../../../actionCreators/users';
 import User from './User';
 import '../../../css/Users.css';
 
@@ -35,4 +38,31 @@ UserList.propTypes = {
   toggleCanVote: PropTypes.func.isRequired,
 };
 
+const mapStateToProps = (state) => {
+  // Use Fuse for fuzzy-search.
+  const fuse = new Fuse(state.users, {
+    threshold: 0.6,
+    location: 0,
+    distance: 100,
+    maxPatternLength: 32,
+    minMatchCharLength: 1,
+    keys: ['name'],
+  });
+
+  return {
+    // Show all users if there is no filter in the box.
+    users: state.userFilter ? fuse.search(state.userFilter) : state.users,
+  };
+};
+
+const mapDispatchToProps = dispatch => ({
+  toggleCanVote: (id) => {
+    dispatch(toggleCanVote(id));
+  },
+});
+
 export default UserList;
+export const UserListContainer = connect(
+  mapStateToProps,
+  mapDispatchToProps,
+)(UserList);
