+++ client/src/components/App/VotingMenu.js
@@ -1,4 +1,8 @@
 import React from 'react';
+import { connect } from 'react-redux';
+import { sendVote } from '../../actionCreators/issues';
+import getShuffledAlternatives from '../../selectors/getShuffledAlternatives';
+import { getIssueId, getIssueKey } from '../../selectors/issues';
 import Alternatives from '../Alternatives';
 import Button from '../Button';
 import '../../css/VotingMenu.css';
@@ -75,4 +79,32 @@ VotingMenu.propTypes = {
   voterKey: React.PropTypes.number,
 };
 
+const mapStateToProps = state => ({
+  // Alternatives are shuffled as an attempt to prevent peeking over shoulders
+  // to figure out what another person has voted for. This scramble needs
+  // to be syncronized between LiveVoteCount and VoteHandler, so we take
+  // advantage of the memoizing provided by reselect. This keeps the
+  // scrambles in sync and avoids rescrambling unless the
+  // available alternatives are changed.
+  alternatives: getShuffledAlternatives(state),
+
+  votes: getIssueKey(state, 'votes', []),
+
+  // The ID, or undefined, if there is no current issue.
+  issueId: getIssueId(state),
+
+  voterKey: state.voterKey,
+});
+
+const mapDispatchToProps = dispatch => ({
+  handleVote: (id, alternative, voter) => {
+    dispatch(sendVote(id, alternative, voter));
+  },
+});
+
+
 export default VotingMenu;
+export const VotingMenuContainer = connect(
+  mapStateToProps,
+  mapDispatchToProps,
+)(VotingMenu);
