+++ client/src/components/IssueStatus.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import LiveVoteCount from '../containers/LiveVoteCount';
+import { VoteStatusContainer } from './VoteStatus';
 import '../css/IssueStatus.css';
 
 const IssueStatus = () => (
@@ -7,7 +7,7 @@ const IssueStatus = () => (
     <h2 className="IssueStatus-title">
       Foreløpig status
     </h2>
-    <LiveVoteCount />
+    <VoteStatusContainer />
   </div>
 );
