+++ client/src/index.js
@@ -5,10 +5,10 @@ import { Router, Route, browserHistory, IndexRoute } from 'react-router';
 import { applyMiddleware, createStore } from 'redux';
 import IO from 'socket.io-client';
 import createSocketIoMiddleware from 'redux-socket.io';
-import AdminPanelContainer from './containers/AdminPanelContainer';
-import AppContainer from './containers/AppContainer';
+import { AdminPanelContainer } from './components/Admin/AdminPanel';
+import { AppContainer } from './components/App/';
 import AdminHome from './components/Admin/Home';
-import IssueFormContainer from './containers/IssueFormContainer';
+import { IssueFormContainer } from './components/Admin/IssueForm';
 import NotFound from './components/NotFound';
 import Users from './components/Admin/Users';
 import votingApp from './reducers';
